full-requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.4
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
astroid==3.0.3
async-timeout==4.0.3
attrs==23.1.0
autocommand==2.2.2
babel==2.17.0
basicsr==1.4.2
beautifulsoup4==4.12.2
bidict==0.22.1
bitsandbytes==0.35.4
blendmodes==2022
blinker==1.6.2
blis==0.7.10
boltons==23.0.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.6
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
colorama==0.4.6
comtypes==1.2.0
confection==0.1.1
contourpy==1.1.0
controlnet-aux==0.0.3
cssselect2==0.7.0
cycler==0.11.0
cymem==2.0.7
dadaptation==3.1
dataclasses-json==0.5.14
deprecation==2.1.0
diffusers==0.16.1
dill==0.3.7
discord-webhook==1.1.0
discord.py==2.5.2
Django==4.2.5
einops==0.4.1
elastic-transport==8.4.0
elasticsearch==8.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
facexlib==0.3.0
faiss-cpu==1.7.4
fastapi==0.94.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.2
filterpy==1.4.5
Flask==2.2.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
flaskwebgui==0.3.5
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.30
google-api-core==1.34.0
google-api-python-client==1.12.8
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
gradio==3.32.0
gradio_client==0.4.0
greenlet==2.0.2
grpcio==1.57.0
gTTS==2.5.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
inflect==7.0.0
inflection==0.5.1
iopath==0.1.9
irc==20.1.0
isort==5.13.2
itsdangerous==2.1.2
jaraco.collections==4.3.0
jaraco.context==4.3.0
jaraco.functools==3.8.1
jaraco.logging==3.1.2
jaraco.stream==3.0.3
jaraco.text==3.11.1
Jinja2==3.1.2
joblib==1.3.2
jsonmerge==1.8.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
kornia==0.6.7
lama-cleaner==1.2.3
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langcodes==3.3.0
langsmith==0.0.85
lark==1.1.2
lazy_loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lion-pytorch==0.1.2
llvmlite==0.40.1
lmdb==1.4.1
logging==0.4.9.6
loguru==0.7.0
lpips==0.1.4
lxml==4.9.3
manifest-ml==0.0.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.3
more-itertools==10.1.0
mpmath==1.3.0
mpyq==0.2.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.5
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
open-clip-torch==2.20.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
orjson==3.9.5
packaging==23.2
pandas==2.0.3
pathy==0.10.2
pefile==2023.2.7
piexif==1.1.3
Pillow==9.5.0
platformdirs==3.10.0
playsound==1.3.0
portalocker==2.7.0
preshed==3.0.8
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydeck==0.8.0
pydub==0.25.1
pygame==2.5.0
Pygments==2.16.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pylint==3.0.3
Pympler==1.0.1
PyMuPDF==1.25.3
PyMySQL==1.1.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-engineio==4.6.0
python-multipart==0.0.6
python-socketio==5.8.0
pytorch-lightning==1.9.4
pyttsx3==2.90
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
realesrgan==0.3.0
redis==4.6.0
referencing==0.30.2
regex==2023.8.8
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
sc2reader==1.8.0
scikit-image==0.20.0
scipy==1.11.2
seaborn==0.12.2
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.1.99
simple-websocket==0.10.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spawningtool @ git+https://github.com/StoicLoofah/spawningtool.git@7566f458b6b40cbdd58ee5cc8794c5b9adde71eb
SpeechRecognition==3.10.4
SQLAlchemy==1.4.49
sqlitedict==2.1.0
sqlparse==0.4.4
srsly==2.4.7
starlette==0.26.1
streamlit==1.25.0
svglib==1.5.1
sympy==1.12
tabulate==0.9.0
tb-nightly==2.15.0a20230820
tempora==5.5.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thinc==8.1.11
tifffile==2023.8.12
tiktoken==0.4.0
timm==0.6.7
tinycss2==1.2.1
tkcalendar==1.6.1
tokenizers==0.13.3
tomesd==0.1.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.0
torch==2.0.1+cu118
torchdiffeq==0.2.3
torchmetrics==1.0.3
torchsde==0.2.5
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.65.0
trampoline==0.1.2
transformers==4.25.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
ultralytics==8.0.158
uritemplate==3.0.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.21.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.7
whichcraft==0.6.1
wikipedia==1.4.0
win32-setctime==1.1.0
wsproto==1.2.0
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
